Add some coloring to Lintje usage example
Add custom coloring to the Lintje usage example on the homepage and docs
index. Move it to a partial so I don't need to maintain it in two
locations.

It's not exactly the colors I want. In my terminal theme the underline
is orange, but the contrast doesn't work in this light theme.