Add some graphics for favicon and social media
This is hopefully not the final design, but it's better than nothing.

I added some poor mans digest at the end of the favicon and other image
paths so that they hopefully expire faster if they change.
I couldn't get it to work with `asset_path`, I got an error I don't have
time to look into now.