Split rules up into its own collection
Rather than add the rule descriptions to a file that grows longer with
every rule, split each rule into its own file in the new rules
collection.

Then generate the rule pages based on those files.

This also allows me to create a list of all rules on the index page that
will always be kept in sync.