Move mobile docs navigation to main navigation
I used a temporarily solution before by rendering the docs navigation
only on the docs pages. This is inconvenient if you want to go to a
docs page directly from the home page.

It also split the navigation in two places, making it unclear where what
could be found.

I've moved the docs navigation to the main navigation, also mobile,
behind the hamburger menu.

I've had to render two different navigations, because I want to have a
different order of items on mobile. This includes putting the search at
the top and the docs links at the bottom.