Describe Git hooks better
Don't recommend the more complicated version of the `commit-msg` hook to
not fail the commit creation. The `post-commit` hook is better suited
for this and requires no error handling.

Lintje requires no arguments for this Git hook. It will by default
validate the last commit, which is the just created commit.